fixed pointer to relation invalid after refresh
Fixed case, when was obtained pointer to the relationship QVector or
std::optional and after the BaseModel::refresh was called, the pointer
was invalidated.
The reason was the QHash type for m_relations data member and implicit
sharing.

The m_relations type was changed to the std::unordered_map to prevent
this behavior.

Also during the BaseModel::load call will be replaced only loaded
relations, all other relations will be untouched.

 - code responsible for creating visitor stores and calling visitors
   was extracted to private methods