get rid of Model::relationVisitor() in models
I have finally found a way how to get rid of the
Model::relationVisitor() in models and also get rid of the std::any
in the Model::u_relations hash.

Everything was completely rewritten, Model::u_relations now accepts
RelationVisitor instead of the std::any as the value which is
std::function<void(BaseRelationStore &)> and the BaseRelationStore
has defined an operator(), so mappings from relation name to the
relation methods are defined like this:

    QHash<QString, RelationVisitor> u_relations {
        {"roles", [](auto &v) { v(&User::roles); }},
    };

This is the nicer and shortest syntax that I came up with. I think
it is fire 🔥🚀

I have also updated all of the models.

From my point of view, this is the biggest success of this project,
because I tried to accomplish this almost half a year, from the
beginning of the project and I could not find the correct way to do it.

I tried every few months to accomplish this and finally, I did it 👏🙌.