removed methods for inserting/updating relations 🚀
As a consequence of the previous few commits, I can remove all the
virtual methods for Inserting / Updating related models.

This is possible because all Model's relation methods are returning
actual relation types instead of base Relation.

I wanted to achieve this, this was the primary target of my changes.