enhanced autotests
Tests don't crash when environment variables for connection/s are
not defined.

The TestUtils::Database class is designed so that a
Database::createConnections/createConnection methods can be called only
once. You can pass connection name/s to these methods and they create
TinyORM database connections.
Only those connections will be created, for which are environment
variables defined correctly.
Tests don't fail but are skipped when a connection is not available.