qmake added Windows RC and manifest files
Windows RC and manifest files are compiled and embedded only on Windows
and when a shared library is built.
Also removed TinyOrm_mingw.rc file, RC file for MinGW is injected
through the main TinyOrm.rc.in file, I had to do it this way because is
not possible to pass more RC files by qmake's RC_FILE variable, cmake
can do this without problems.

 - bugfix FILEFLAGS, FILEFLAGSMASK for debug configuration was not set