divided and allocated commands on the heap
Commands was extracted from the ColumnDefinition and has been created
own struct for every or very similar commands. They are now allocated
on the heap to save space because sizeof(ColumnDefinition) was 736 and
that is too much, eg if schema would contain 100 columns it would take
73KB.
I have decided not to use polymorphic commands, I wanted to use
designated initializers with aggregates, the consequence of this is
usage of reinterpret_cast() :/, but it works great.

 - renamed commands from XxCommandDefinition to XxCommand
 - get rid of a data member name collisions (eg index_, from_, ...)
   in the column and command definitions, consequence of division
   column and commands