added unix gcc/clang support for schema builder
I had to reject designated initializers with commands because of gcc
throws -Wmissing-field-initializers warning with aggregates that have
a base class even empty base class, so I had to use classic aggregate
initialization with commands, it's not a big deal as all data members
of commands have to be initialized anyway as command aggregate classes
are tailor-made for a particular commands. I didn't want to suppress
this warning around every addCommand() method call.

 - mainly default init. of data members
 - some shadow/overlapping variable names in lambdas
 - missing SHAREDLIB_EXPORT for SchemmaGrammar