added table name guesser for make:migration ✨🚀
make:migration command is able to guess the table name and creation
status from the passed migration name on the command line.

If the migration name begins with the create_ string then the create
migration stub will be used.

If the migration name contains the (to|from|in)_ pattern then the update
migration stub will be used.

In all other cases an empty migration stub will be used.