added integrate for zsh shell
It's design to work on Ubuntu, Manjaro, Arch, and Gentoo, if any of
these folders exist /usr/[local/]share/zsh/site-functions then write
the _tom completion file into it, if don't then try to create
the /usr/local/share/zsh/site-functions folder and write to it.
These folders should be on the zsh's $fpath on all above described
OS-es.