added Soft Deleting support 🤯🙌
In addition to actually removing records from your database, TinyORM
can also "soft delete" models. When models are soft deleted, they are
not actually removed from your database. Instead, a deleted_at attribute
is set on the model indicating the date and time at which the model was
"deleted".

 - added docs
 - added all proxies
 - added complete functional tests and a few unit tests
 - carefully verified all possible scenarios because it changes
   practically all queries 🤯

Others:

 - also added from() to all proxies
 - bugfix in seeders, bad ID in the PostgreSQL primary key (sequence)