fixed handling of timestamp columns on Pivots
Now, the timestamp column names on the Pivot tables can be different
than timestamp columns on the Parent model.
They can be set by the withTimestamps() method. If they are not set
by this method they still be inferred from attributes and the timestamp
column names will be inferred from the Parent model.

Also the usesTimestamps() logic was enhanced on the Pivot models,
both timestamp column names (createdAt and updatedAt) have to be set
in the attributes, to set usesTimestamps() to true.