tom enhanced detection of paths for make: commands
The make:migration/seeder/model commands try to guess a path based
on the pwd and the defaultPath (can be relative or absolute; it can be
migrations, seeders, or models default path).

If the defaultPath is relative and the pwd is somewhere inside this
defaultPath, then it appends the rest of the path from the defaultPath
to the pwd, eg. the defaultPath = "aa/bb/cc/dd" and
the pwd = "E:/<some path>/aa/bb", then the guessed result path will be
"E:/<some path>/aa/bb/cc/dd", the "cc/dd" was appended.

This is the prime case how the intelligent logic can corrupt a nice
code, many people hate this type of logic, but whatever, it works
great, the logic is outsourced to the tom utils class, so it's not so
bad. 😵‍💫🙃