cmake added support for cmp0141 with ccache
Added support for the MSVC debug information format flags selected by
an abstraction added in the CMake 3.25 if the compiler launcher is
the ccache.

In the CMake >=3.25 the CMAKE_MSVC_DEBUG_INFORMATION_FORMAT will be
set to the "$<$<CONFIG:Debug,RelWithDebInfo>:Embedded>" if the compiler
launcher is the ccache.

In the CMake <3.25 the old replace /Zi by /Z7 by the build config type
in the CMAKE_C/CXX_FLAGS_DEBUG/RELWITHDEBINFO CMake flags.