workflows disabled ccache compression
Also updated max_size for all actions, with this formula:
<initial_size> * 3 + 100

The <initial_size> is a size after compression, I have recompiled
with all supported compilers and platforms and checked this initial
size.

The problem with the ccache misses was disabled ccache compression, so
the problem was a low ccache max_size.