workflows bugfix empty file with tee
The problem was the cat with the tee (w/o --append) on the same file,
the tee command is sometimes faster than the cat command, so
it truncates the file content before the cat can read it.

The solution was the sed --in-place, so no need to use the cat and tee.