added SSL-related PostgreSQL configuration
Added a new SSL-related configuration options to the top-level
configuration level for PostgreSQL database. They are sslmode, sslcert,
sslkey, and sslrootcert.

These options are copied from the top-level configuration to the options
option hash during configuration parsing
by the PostgresConfigurationParser.

If both are defined, then the top-level option overwrites a option
in the 'options' hash.

 - also added constants
 - functional test
 - updated docs