schema used modifier to set timestamp default
Used TinyORM modifiers to set a default value and for the MySQL also
"on update" value for the datetime and timestamp column types.

Previously, the SQL for a default value and for the "on update" was
generated right away, now, the ColumnDefinition instance is modified
and later in the SQL generation process will be this clauses generated.

 - updated tests, the order of the null value and default values were
   switched