added Appending Values To JSON support 🕺
These Appends are appended to the serialized models during the toMap(),
toVector(), or toJson() methods call.

Appends names are and must be mapped using the u_mutators hash map
(in the same way like u_relations) because c++ doesn't support advanced
reflection.

Appends are mapped to methods (can be protected) that return
the Casts::Attribute. This Casts::Attribute is a wrapper to return
a new attribute value.

Appends accept u_visible and u_hidden.

For the Appends that return QDate or QDateTime are the counterpart
serializeDate() or serializeDateTime() methods called.

 - added unit tests