Include exception headers (silverqx/TinyORM#16)
The querybuilders' unit tests fail to compile due to some Exception
classes being undefined if ORM compilation is disabled via cmake.

The undefined exceptions are actually being built as part of the
querybuilder. However, the unit tests deep-end on models/user.hpp to
include the header files for the exception classes they are using,
but the corresponding include directive is skipped for builds that
disable ORM.