qmake added tiny_dotenv feature
It allows to load the .env and .env.$$TINY_DOTENV_PLATFORM files. These
files are included as early as possible so they can setup or affect
other configuration steps.

The .env file is included always and the .env.$$TINY_DOTENV_PLATFORM
file is included only for the current platform.
The $$TINY_DOTENV_PLATFORM qmake variable is auto-detected from
the $$QMAKE_PLATFORM.

The opposite thing to this feature are conf.pri files in which you can
override the entire configuration because they are included as late
as possible.