tools deploy added updating vcpkg port-version
It operates on two vcpkg ports, on the tinyorm and tinyorm-qt5 ports.

When the TinyORM library version was bumped then it removes
the port-version field from the vcpkg.json files.
If the version was not bumped then it ask if you want to bump
the port-version field, you can select for which port you want to bump
the port-version number.