workflows bugfix used github.sha
With the github.ref_name the job can fail if there are new commits
in the current branch because in this case the SHA512 will not match.

It's happening eg. during deploy.ps1 when the vcpkg REF and SHA512 is
committed right after the bump version number commit.

Using the github.sha is safer anyway.

[skip ci]