drivers bugfix driver() vs driverWeak()
As the TinyDrivers library uses smart pointers everywhere it stores
the SqlDriver as std::shared_ptr<>/weak_ptr<> and because of this
we end up with API compatibility problems, exactly
the exposed SqlDriver *driver() method was problematic.

This commit solves this problem. The idea was to provide compatible API
related to this driver() method but also to still be able use smart
pointers internally.