drivers enhanced loading SQL driver at runtime
If loading the SQL driver shared library at runtime from all standard
paths failed then load the SQL driver library from the build tree. Load
it only if the TinyDrivers library location is in the build tree. This
ensures that it will not be loaded from the build tree, e.g. after
installation because it is non-standard behavior.