workflows ran others workflows at end
The reason for this is the concurrency:group: can contain only two
in_progress workflows, one will be in_progress and the second will be
queued (waiting until the first finish), and all others will be
canceled.

Also, I can't test this commit until it will be merged to the main
branch.