cmake check minimum required Qt version
Determine whether the minimum Qt version was satisfied using and set the
internal cache variable TINY_QT_VERSION.

Command used to obtain a Qt version: "${QT_QMAKE_EXECUTABLE}" -query
QT_VERSION This check is needed because eg. QtCreator sets the
QT_QMAKE_EXECUTABLE based on the selected KIT, but there can be other Qt
versions on the system/user path so even if the QT_QMAKE_EXECUTABLE
contains qmake from eg. Qt v5.15 then the find_package() function is
still able to find the correct Qt version, so this check is specifically
for the QtCreator and informs about wrongly select KIT.

If the QT_QMAKE_EXECUTABLE command can't be executed or it returns a
non-zero exit code then continue a normal execution and leave the
decision logic up to the find_package() function.